Extracted slack response for search results into separate function
In order to achieve a consistent look-and-feel for responses, extracted slack response for search results into a separate function and added to a new file richMessageHelper.py. This will facilitate reuse of the function by other intents to display search results

Added new dependency pyngrok. On running of flask app, if RUN_NGROK ==
True, automatically launch a ngrok instance and print and get the public url. Please use the
url printed in the console to update DialogFlow Fulfillment Webhook url.